Splitting into multiple modules
